1. Abstracted the Window/Model design, into a SceneManager/Scene design. 2. Consolidated OpenGL specific code into game.graphics.py
