+++ game/graphics.py
@@ -1,0 +1,108 @@
+#!/usr/bin/python3
+# -*- coding: utf-8 -*-
+
+"""
+ ________                                        ______                       ______     __
+|        \                                      /      \                     /      \   |  \
+ \$$$$$$$$______    ______    ______   ______  |  $$$$$$\  ______   ______  |  $$$$$$\ _| $$_
+   | $$  /      \  /      \  /      \ |      \ | $$   \$$ /      \ |      \ | $$_  \$$|   $$ \
+   | $$ |  $$$$$$\|  $$$$$$\|  $$$$$$\ \$$$$$$\| $$      |  $$$$$$\ \$$$$$$\| $$ \     \$$$$$$
+   | $$ | $$    $$| $$   \$$| $$   \$$/      $$| $$   __ | $$   \$$/      $$| $$$$      | $$ __
+   | $$ | $$$$$$$$| $$      | $$     |  $$$$$$$| $$__/  \| $$     |  $$$$$$$| $$        | $$|  \
+   | $$  \$$     \| $$      | $$      \$$    $$ \$$    $$| $$      \$$    $$| $$         \$$  $$
+    \$$   \$$$$$$$ \$$       \$$       \$$$$$$$  \$$$$$$  \$$       \$$$$$$$ \$$          \$$$$
+
+
+Copyright (C) 2013 Michael Fogleman
+Copyright (C) 2018 Stefano Peris <xenon77.dev@gmail.com>
+
+Github repository: <https://github.com/XenonCoder/Terracraft>
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+
+from pyglet.gl import *
+
+from .config import *
+
+
+def setup_fog():
+    """ Configure the OpenGL fog properties.
+
+    """
+    # Enable fog. Fog "blends a fog color with each rasterized pixel fragment's
+    # post-texturing color."
+    glEnable(GL_FOG)
+    # Set the fog color.
+    glFogfv(GL_FOG_COLOR, (GLfloat * 4)(0.5, 0.69, 1.0, 1))
+    # Say we have no preference between rendering speed and quality.
+    glHint(GL_FOG_HINT, GL_DONT_CARE)
+    # Specify the equation used to compute the blending factor.
+    glFogi(GL_FOG_MODE, GL_LINEAR)
+    # How close and far away fog starts and ends. The closer the start and end,
+    # the denser the fog in the fog range.
+    glFogf(GL_FOG_START, FOG_START)
+    glFogf(GL_FOG_END, FOG_END)
+
+
+def setup_opengl():
+    """ Basic OpenGL configuration.
+
+    """
+    # Set the color of "clear", i.e. the sky, in rgba.
+    glClearColor(0.5, 0.69, 1.0, 1)
+    # Enable culling (not rendering) of back-facing facets -- facets that aren't
+    # visible to you.
+    glEnable(GL_CULL_FACE)
+    # Set the texture minification/magnification function to GL_NEAREST (nearest
+    # in Manhattan distance) to the specified texture coordinates. GL_NEAREST
+    # "is generally faster than GL_LINEAR, but it can produce textured images
+    # with sharper edges because the transition between texture elements is not
+    # as smooth."
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST)
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST)
+    setup_fog()
+
+
+def set_2d(window_size):
+    """ Configure OpenGL to draw in 2d.
+
+    """
+    width, height = window_size
+    glDisable(GL_DEPTH_TEST)
+    glViewport(0, 0, width, height)
+    glMatrixMode(GL_PROJECTION)
+    glLoadIdentity()
+    glOrtho(0, width, 0, height, -1, 1)
+    glMatrixMode(GL_MODELVIEW)
+    glLoadIdentity()
+
+
+def set_3d(window_size, rotation=(0, 0), position=(0, 0, 0)):
+    """ Configure OpenGL to draw in 3d.
+
+    """
+    width, height = window_size
+    glEnable(GL_DEPTH_TEST)
+    glViewport(0, 0, width, height)
+    glMatrixMode(GL_PROJECTION)
+    glLoadIdentity()
+    gluPerspective(65.0, width / float(height), 0.1, 60.0)
+    glMatrixMode(GL_MODELVIEW)
+    glLoadIdentity()
+    x, y = rotation
+    glRotatef(x, 0, 1, 0)
+    glRotatef(-y, math.cos(math.radians(x)), 0, math.sin(math.radians(x)))
+    x, y, z = position
+    glTranslatef(-x, -y, -z)
