+++ game/scenes.py
@@ -2,12 +2,12 @@
 
 """
  ________                                        ______                       ______     __
-|        \                                      /      \                     /      \   |  \ 
+|        \                                      /      \                     /      \   |  \
  \$$$$$$$$______    ______    ______   ______  |  $$$$$$\  ______   ______  |  $$$$$$\ _| $$_
-   | $$  /      \  /      \  /      \ |      \ | $$   \$$ /      \ |      \ | $$_  \$$|   $$ \ 
+   | $$  /      \  /      \  /      \ |      \ | $$   \$$ /      \ |      \ | $$_  \$$|   $$ \
    | $$ |  $$$$$$\|  $$$$$$\|  $$$$$$\ \$$$$$$\| $$      |  $$$$$$\ \$$$$$$\| $$ \     \$$$$$$
    | $$ | $$    $$| $$   \$$| $$   \$$/      $$| $$   __ | $$   \$$/      $$| $$$$      | $$ __
-   | $$ | $$$$$$$$| $$      | $$     |  $$$$$$$| $$__/  \| $$     |  $$$$$$$| $$        | $$|  \ 
+   | $$ | $$$$$$$$| $$      | $$     |  $$$$$$$| $$__/  \| $$     |  $$$$$$$| $$        | $$|  \
    | $$  \$$     \| $$      | $$      \$$    $$ \$$    $$| $$      \$$    $$| $$         \$$  $$
     \$$   \$$$$$$$ \$$       \$$       \$$$$$$$  \$$$$$$  \$$       \$$$$$$$ \$$          \$$$$
 
@@ -31,22 +31,41 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
+import random
+import time
 import pyglet
 
+from collections import deque
+
+from pyglet import image
 from pyglet.gl import *
 from pyglet.window import key, mouse
+from pyglet.graphics import TextureGroup
 
 from .blocks import *
-from .model import Model
-from .vox_cube import cube_vertices
-from .vox_cube import normalize
-from .vox_cube import sectorize
-from .config import *
+from .vox_cube import *
+from .graphics import set_2d, set_3d
+from .savemanager import SaveManager
 
 
-class Window(pyglet.window.Window):
-    def __init__(self, *args, **kwargs):
-        super(Window, self).__init__(*args, **kwargs)
+class Scene:
+    """A base class for all game scenes."""
+
+    def update(self, dt):
+        raise NotImplementedError
+
+
+class MenuScene(Scene):
+    def __init__(self, window):
+        self.window = window
+
+    def update(self, dt):
+        pass
+
+
+class GameScene(Scene):
+    def __init__(self, window):
+        self.window = window
 
         # Whether or not the window exclusively captures the mouse.
         self.exclusive = False
@@ -100,16 +119,15 @@ class Window(pyglet.window.Window):
         self.block = self.inventory[0]
 
         # Convenience list of num keys.
-        self.num_keys = [
-            key._1, key._2, key._3, key._4, key._5,
-            key._6, key._7, key._8, key._9, key._0]
+        self.num_keys = [key._1, key._2, key._3, key._4, key._5,
+                         key._6, key._7, key._8, key._9, key._0]
 
         # Instance of the model that handles the world.
         self.model = Model()
 
         # The label that is displayed in the top left of the canvas.
         self.info_label = pyglet.text.Label('', font_name='Arial', font_size=INFO_LABEL_FONTSIZE,
-                                            x=10, y=self.height - 10, anchor_x='left',
+                                            x=10, y=self.window.height - 10, anchor_x='left',
                                             anchor_y='top',
                                             color=(0, 0, 0, 255))
 
@@ -117,7 +135,7 @@ class Window(pyglet.window.Window):
         self.is_initializing = True
         # Loading screen label displayed in center of canvas.
         self.loading_label = pyglet.text.Label('', font_name='Arial', font_size=50,
-                                               x=self.width // 2, y=self.height // 2,
+                                               x=self.window.width // 2, y=self.window.height // 2,
                                                anchor_x='center', anchor_y='center',
                                                color=(0, 0, 0, 255))
 
@@ -126,7 +144,7 @@ class Window(pyglet.window.Window):
         the game will ignore the mouse.
 
         """
-        super(Window, self).set_exclusive_mouse(exclusive)
+        self.window.set_exclusive_mouse(exclusive)
         self.exclusive = exclusive
 
     def get_sight_vector(self):
@@ -144,7 +162,7 @@ class Window(pyglet.window.Window):
         dy = math.sin(math.radians(y))
         dx = math.cos(math.radians(x - 90)) * m
         dz = math.sin(math.radians(x - 90)) * m
-        return (dx, dy, dz)
+        return dx, dy, dz
 
     def get_motion_vector(self):
         """ Returns the current motion vector indicating the velocity of the
@@ -187,7 +205,7 @@ class Window(pyglet.window.Window):
             dy = 0.0
             dx = 0.0
             dz = 0.0
-        return (dx, dy, dz)
+        return dx, dy, dz
 
     def update(self, dt):
         """ This method is scheduled to be called repeatedly by the pyglet
@@ -375,8 +393,8 @@ class Window(pyglet.window.Window):
             elif self.dy == 0:
                 self.dy = JUMP_SPEED
         elif symbol == key.ESCAPE:
-            self.set_exclusive_mouse(False)
-            # exit()
+                self.set_exclusive_mouse(False)
+                return pyglet.event.EVENT_HANDLED
         elif symbol == key.TAB:
             self.flying = not self.flying
         elif symbol == key.F1:
@@ -427,58 +445,27 @@ class Window(pyglet.window.Window):
         # reticle
         if self.reticle:
             self.reticle.delete()
-        x, y = self.width // 2, self.height // 2
+        x, y = self.window.width // 2, self.window.height // 2
         n = 10
         self.reticle = pyglet.graphics.vertex_list(
             4, ('v2i', (x - n, y, x + n, y, x, y - n, x, y + n))
         )
 
-    def set_2d(self):
-        """ Configure OpenGL to draw in 2d.
-
-        """
-        width, height = self.get_size()
-        glDisable(GL_DEPTH_TEST)
-        glViewport(0, 0, width, height)
-        glMatrixMode(GL_PROJECTION)
-        glLoadIdentity()
-        glOrtho(0, width, 0, height, -1, 1)
-        glMatrixMode(GL_MODELVIEW)
-        glLoadIdentity()
-
-    def set_3d(self):
-        """ Configure OpenGL to draw in 3d.
-
-        """
-        width, height = self.get_size()
-        glEnable(GL_DEPTH_TEST)
-        glViewport(0, 0, width, height)
-        glMatrixMode(GL_PROJECTION)
-        glLoadIdentity()
-        gluPerspective(65.0, width / float(height), 0.1, 60.0)
-        glMatrixMode(GL_MODELVIEW)
-        glLoadIdentity()
-        x, y = self.rotation
-        glRotatef(x, 0, 1, 0)
-        glRotatef(-y, math.cos(math.radians(x)), 0, math.sin(math.radians(x)))
-        x, y, z = self.position
-        glTranslatef(-x, -y, -z)
-
     def on_draw(self):
         """ Called by pyglet to draw the canvas.
-        
+
         """
-        self.clear()
+        self.window.clear()
         if not self.is_initializing:
-            self.set_3d()
+            set_3d(self.window.get_size(), self.rotation, self.position)
             glColor3d(1, 1, 1)
             self.model.batch.draw()
             if self.toggleGui:
                 self.draw_focused_block()
-                self.set_2d()
+                set_2d(self.window.get_size())
                 if self.toggleLabel:
                     self.draw_label()
-                    self.set_2d()
+                    set_2d(self.window.get_size())
                 self.draw_reticle()
         # Loading screen
         """ the "self.set_2d()" and "self.draw_label()" functions
@@ -486,7 +473,7 @@ class Window(pyglet.window.Window):
             for draw_focused_block, and draw_label (label for info fps/coords).
         """
         if self.is_initializing:
-            self.set_2d()
+            set_2d(self.window.get_size())
             self.draw_label()
             self.loading_label.delete()
             self.is_initializing = False
@@ -514,7 +501,7 @@ class Window(pyglet.window.Window):
             x, y, z = self.position
             self.info_label.text = 'FPS = [%02d] : COORDS = [%.2f, %.2f, %.2f] : %d / %d' % (
                 pyglet.clock.get_fps(), x, y, z,
-                len(self.model._shown), len(self.model.world))
+                self.model.currently_shown, len(self.model.world))
             self.info_label.draw()
         else:
             # Only draw the loading screen during the first draw loop.
@@ -528,3 +515,322 @@ class Window(pyglet.window.Window):
         if not self.is_initializing:
             glColor3d(0, 0, 0)
             self.reticle.draw(GL_LINES)
+
+
+class Model(object):
+    def __init__(self):
+        # A Batch is a collection of vertex lists for batched rendering.
+        self.batch = pyglet.graphics.Batch()
+
+        # A TextureGroup manages an OpenGL texture.
+        self.group = TextureGroup(image.load(TEXTURE_PATH).get_texture())
+
+        # A mapping from position to the texture of the block at that position.
+        # This defines all the blocks that are currently in the world.
+        self.world = {}
+
+        # Same mapping as `world` but only contains blocks that are shown.
+        self.shown = {}
+
+        # Mapping from position to a pyglet `VertextList` for all shown blocks.
+        self._shown = {}
+
+        # Mapping from sector to a list of positions inside that sector.
+        self.sectors = {}
+
+        # Simple function queue implementation. The queue is populated with
+        # _show_block() and _hide_block() calls
+        self.queue = deque()
+
+        # A module to save and load the world
+        self.save_manager = SaveManager()
+
+        self._initialize()
+
+    @property
+    def currently_shown(self):
+        return len(self._shown)
+
+    def _initialize(self):
+        """ Initialize the world by placing all the blocks.
+
+        Blocks will be loaded from save if available.
+        Otherwise, a new map will be generated.
+
+        """
+        if self.save_manager.has_save_game():
+            # Returns False if unable to load the save
+            has_save = self.save_manager.load_world(self)
+        else:
+            has_save = False
+
+        if not has_save:
+            n = 80  # 1/2 width and height of world
+            s = 1  # step size
+            y = 0  # initial y height
+            for x in range(-n, n + 1, s):
+                for z in range(-n, n + 1, s):
+                    # create a layer stone an DIRT_WITH_GRASS everywhere.
+                    self.add_block((x, y - 2, z), DIRT_WITH_GRASS, immediate=False)
+                    self.add_block((x, y - 3, z), BEDSTONE, immediate=False)
+                    if x in (-n, n) or z in (-n, n):
+                        # create outer walls.
+                        for dy in range(-2, 3):
+                            self.add_block((x, y + dy, z), BEDSTONE, immediate=False)
+
+            # generate the hills randomly
+            o = n - 10
+            for _ in range(120):
+                a = random.randint(-o, o)  # x position of the hill
+                b = random.randint(-o, o)  # z position of the hill
+                c = -1  # base of the hill
+                h = random.randint(1, 6)  # height of the hill
+                s = random.randint(4, 8)  # 2 * s is the side length of the hill
+                d = 1  # how quickly to taper off the hills
+                t = random.choice([DIRT_WITH_GRASS, SNOW, SAND])
+                for y in range(c, c + h):
+                    for x in range(a - s, a + s + 1):
+                        for z in range(b - s, b + s + 1):
+                            if (x - a) ** 2 + (z - b) ** 2 > (s + 1) ** 2:
+                                continue
+                            if (x - 0) ** 2 + (z - 0) ** 2 < 5 ** 2:  # 6 = flat map
+                                continue
+                            self.add_block((x, y, z), t, immediate=False)
+                    s -= d  # decrement side lenth so hills taper off
+
+    def hit_test(self, position, vector, max_distance=NODE_SELECTOR):
+        """ Line of sight search from current position. If a block is
+        intersected it is returned, along with the block previously in the line
+        of sight. If no block is found, return None, None.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position to check visibility from.
+        vector : tuple of len 3
+            The line of sight vector.
+        max_distance : int
+            How many blocks away to search for a hit.
+
+        """
+        m = 8
+        x, y, z = position
+        dx, dy, dz = vector
+        previous = None
+        for _ in range(max_distance * m):
+            key = normalize((x, y, z))
+            if key != previous and key in self.world:
+                return key, previous
+            previous = key
+            x, y, z = x + dx / m, y + dy / m, z + dz / m
+        return None, None
+
+    def exposed(self, position):
+        """ Returns False if given `position` is surrounded on all 6 sides by
+        blocks, True otherwise.
+
+        """
+        x, y, z = position
+        for dx, dy, dz in FACES:
+            if (x + dx, y + dy, z + dz) not in self.world:
+                return True
+        return False
+
+    def add_block(self, position, texture, immediate=True):
+        """ Add a block with the given `texture` and `position` to the world.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to add.
+        texture : list of len 3
+            The coordinates of the texture squares. Use `tex_coords()` to
+            generate.
+        immediate : bool
+            Whether or not to draw the block immediately.
+
+        """
+        if position in self.world:
+            self.remove_block(position, immediate)
+        self.world[position] = texture
+        self.sectors.setdefault(sectorize(position), []).append(position)
+        if immediate:
+            if self.exposed(position):
+                self.show_block(position)
+            self.check_neighbors(position)
+
+    def remove_block(self, position, immediate=True):
+        """ Remove the block at the given `position`.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to remove.
+        immediate : bool
+            Whether or not to immediately remove block from canvas.
+
+        """
+        del self.world[position]
+        self.sectors[sectorize(position)].remove(position)
+        if immediate:
+            if position in self.shown:
+                self.hide_block(position)
+            self.check_neighbors(position)
+
+    def check_neighbors(self, position):
+        """ Check all blocks surrounding `position` and ensure their visual
+        state is current. This means hiding blocks that are not exposed and
+        ensuring that all exposed blocks are shown. Usually used after a block
+        is added or removed.
+
+        """
+        x, y, z = position
+        for dx, dy, dz in FACES:
+            key = (x + dx, y + dy, z + dz)
+            if key not in self.world:
+                continue
+            if self.exposed(key):
+                if key not in self.shown:
+                    self.show_block(key)
+            else:
+                if key in self.shown:
+                    self.hide_block(key)
+
+    def show_block(self, position, immediate=True):
+        """ Show the block at the given `position`. This method assumes the
+        block has already been added with add_block()
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to show.
+        immediate : bool
+            Whether or not to show the block immediately.
+
+        """
+        block = self.world[position]
+        self.shown[position] = block
+        if immediate:
+            self._show_block(position, block)
+        else:
+            self._enqueue(self._show_block, position, block)
+
+    def _show_block(self, position, block):
+        """ Private implementation of the `show_block()` method.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to show.
+        block : Block instance
+            An instance of the Block class
+
+        """
+        x, y, z = position
+        vertex_data = cube_vertices(x, y, z, 0.5)
+        # texture_data = list(texture)
+        # create vertex list
+        # FIXME Maybe `add_indexed()` should be used instead
+        self._shown[position] = self.batch.add(24, GL_QUADS, self.group,
+                                               ('v3f/static', vertex_data),
+                                               ('t2f/static', block.tex_coords))
+
+    def hide_block(self, position, immediate=True):
+        """ Hide the block at the given `position`. Hiding does not remove the
+        block from the world.
+
+        Parameters
+        ----------
+        position : tuple of len 3
+            The (x, y, z) position of the block to hide.
+        immediate : bool
+            Whether or not to immediately remove the block from the canvas.
+
+        """
+        self.shown.pop(position)
+        if immediate:
+            self._hide_block(position)
+        else:
+            self._enqueue(self._hide_block, position)
+
+    def _hide_block(self, position):
+        """ Private implementation of the 'hide_block()` method.
+
+        """
+        self._shown.pop(position).delete()
+
+    def show_sector(self, sector):
+        """ Ensure all blocks in the given sector that should be shown are
+        drawn to the canvas.
+
+        """
+        for position in self.sectors.get(sector, []):
+            if position not in self.shown and self.exposed(position):
+                self.show_block(position, False)
+
+    def hide_sector(self, sector):
+        """ Ensure all blocks in the given sector that should be hidden are
+        removed from the canvas.
+
+        """
+        for position in self.sectors.get(sector, []):
+            if position in self.shown:
+                self.hide_block(position, False)
+
+    def change_sectors(self, before, after):
+        """ Move from sector `before` to sector `after`. A sector is a
+        contiguous x, y sub-region of world. Sectors are used to speed up
+        world rendering.
+
+        """
+        before_set = set()
+        after_set = set()
+        pad = 4
+        for dx in range(-pad, pad + 1):
+            for dy in [0]:  # range(-pad, pad + 1):
+                for dz in range(-pad, pad + 1):
+                    if dx ** 2 + dy ** 2 + dz ** 2 > (pad + 1) ** 2:
+                        continue
+                    if before:
+                        x, y, z = before
+                        before_set.add((x + dx, y + dy, z + dz))
+                    if after:
+                        x, y, z = after
+                        after_set.add((x + dx, y + dy, z + dz))
+        show = after_set - before_set
+        hide = before_set - after_set
+        for sector in show:
+            self.show_sector(sector)
+        for sector in hide:
+            self.hide_sector(sector)
+
+    def _enqueue(self, func, *args):
+        """ Add `func` to the internal queue.
+
+        """
+        self.queue.append((func, args))
+
+    def _dequeue(self):
+        """ Pop the top function from the internal queue and call it.
+
+        """
+        func, args = self.queue.popleft()
+        func(*args)
+
+    def process_queue(self):
+        """ Process the entire queue while taking periodic breaks. This allows
+        the game loop to run smoothly. The queue contains calls to
+        _show_block() and _hide_block() so this method should be called if
+        add_block() or remove_block() was called with immediate=False
+
+        """
+        start = time.clock()
+        while self.queue and time.clock() - start < 1.0 / TICKS_PER_SEC:
+            self._dequeue()
+
+    def process_entire_queue(self):
+        """ Process the entire queue with no breaks.
+
+        """
+        while self.queue:
+            self._dequeue()
